Added protection in `end_processing' to pervent a crash when the timer is Void or already destroyed.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@39935 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/build2/utilities/gb_digit_checker.e
@@ -24,7 +24,9 @@ feature -- Basic operations
 		do
 			timer_counter := 0
 			timer_counter_at_last_press := 0
-			timer.destroy
+			if timer /= Void and then not timer.is_destroyed then
+				timer.destroy
+			end
 		end
 
 feature -- Status report
